Finished blog implementation
Lots of changes required: removed News from the model (not required by
the design), added a description attribute to Page, and a markup
attribute to accompany all description attributes (indicating the markup
lagnauge that the description is written in).

In order to accommodate the design of the blog list sitting flush with
the top nav-bar several bits of CSS needed alteration requiring minor
alteration to most other pages. Also incorporate LigatureSymbols for the
blog design.

Updated the comics handler to remove the 'blog' comic from the main
comic list and updated the index page to point blog issues at the
blog_issue handler instead of the usual page handler.