Completely re-worked file-handling
Comic files are now arbitrarily named with filenames stored in the
database. File content alteration is handled by writing to new files
then updating the filename in the database (so requests either see the
old filename, and thus the old file, or the new filename and content). A
background thread clears up files once the new filename is definitely
committed (or clears up left over files in the case of a rollback).

This solves a long-standing issue with renumbering of pages or issues
that it would break the link with the corresponding files on disk.
Deletion of pages and issues now causes later pages and issues to
"shuffle back" closing the hole. Currently there is no facility for
insertion of pages or issues.