Add SQL definition of database
The new database structure handles most of the calculation that we've
been doing with complex SQLAlchemy stuff with views in the database.
Instead-of triggers are used to ensure the views (which sit over the
base tables) remain writeable. Explicit primary/foreign key definitions
will be needed in the ORM as obviously the views won't reflect these but
otherwise they should be mostly drop in replacements for the base
tables.